src/userdata/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
gunicorn==21.2.0
django-environ
djangorestframework
Pillow
numpy
